successful build
seperated all the files for ease of use
