+++ include/awp.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+
+void solo_awp();
