+++ include/blockScoring.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void scoreBlocks(int numberOfBlocks, bool colorsort);
