+++ include/distanceSensor.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+
+float wallDistance(bool shouldPrint, bool useRightSensor);
