+++ include/oneGoal.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+
+void one_goal_left();
+void one_goal_right();
