+++ include/twoGoal.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+
+void two_goal_RIGHT();
+void two_goal_LEFT();
